# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Context-Aware PQC Risk Assessment API
  version: 0.1.0
  license:
    name: Internal Capstone Project
    url: https://github.com/ehdgks0627/crypto-scanner
  description: |
    API contract for the PQC Risk Assessment System.

    `docs/api/openapi.yaml` is the design source of truth. The Django Ninja
    implementation must expose an equivalent `/api/openapi.json`.
servers:
  - url: /api
security:
  - ApiToken: []
...
paths:
...
  /jobs:
    get:
      tags: [Jobs]
      summary: List asynchronous jobs
      description: Lists API-visible asynchronous jobs. `POST /jobs` creates scan jobs; discovery and recompute jobs are created by their domain endpoints.
      operationId: listJobs
      parameters:
        - name: kind
          in: query
          schema:
            $ref: '#/components/schemas/JobKind'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/JobStatus'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Scan job list
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobEnvelopePage'
        '400':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Jobs]
      summary: Start a scan job
      operationId: createScanJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanJobCreate'
            examples:
              default:
                externalValue: ./examples/job-create.json
      responses:
        '202':
          description: Scan job accepted
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobEnvelope'
              examples:
                accepted:
                  externalValue: ./examples/scan-job-accepted.json
        '400':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/Error'
  /jobs/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Jobs]
      summary: Get asynchronous job status
      operationId: getJob
      responses:
        '200':
          description: Scan job status
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobEnvelope'
              examples:
                running:
                  externalValue: ./examples/job-running.json
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /jobs/{id}/cancel:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Jobs]
      summary: Request cancellation for an async job
      operationId: cancelJob
      description: |
        Cancels pending/running scan and discovery jobs. Recompute jobs are cancellable only while PENDING.
        RUNNING scan/discovery jobs set cancel_requested_at and transition to CANCELLED when the worker observes the request.
      responses:
        '202':
          description: Cancellation request accepted
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/JobNotCancellable'
        default:
          $ref: '#/components/responses/Error'
  /jobs/{id}/logs:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Jobs]
      summary: List scan run logs
      operationId: listJobLogs
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Scan run logs
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanRunLogPage'
              examples:
                default:
                  externalValue: ./examples/job-logs.json
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    ApiToken:
      type: apiKey
      in: header
      name: X-API-Token
    BootstrapToken:
      type: apiKey
      in: header
      name: X-Bootstrap-Token
    AgentBearer:
      type: http
      scheme: bearer
  headers:
    XRequestId:
      description: Request trace id. If the client sends `X-Request-Id`, the server echoes it.
      schema:
        type: string
        format: uuid
    NoStore:
      description: Polling responses are not cacheable.
      schema:
        type: string
      example: no-store
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      description: Field name, prefixed with `-` for descending order.
      schema:
        type: string
...
  responses:
    Error:
      description: Error response
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              externalValue: ./examples/error-validation.json
    NotFound:
      description: Resource not found
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    JobNotCancellable:
      description: Job cannot be cancelled in its current state or by kind policy.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            job_not_cancellable:
              externalValue: ./examples/error-job-not-cancellable.json
...
    Unprocessable:
      description: Request is semantically invalid
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Required backend dependency is unavailable.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    JobStatus:
      type: string
      enum: [PENDING, RUNNING, COMPLETED, FAILED, CANCELLED]
    JobKind:
      type: string
      enum: [scan_job, discovery, recompute]
    RunStatus:
      type: string
      enum: [SUCCESS, PARTIAL, TIMEOUT, UNREACHABLE, ERROR, SKIPPED]
...
    ScannerId:
      type: string
      enum:
        - network
        - agent.cert_store
        - agent.pkg_keyring
        - agent.ssh_userkey
        - agent.ssh_config
        - agent.keystore
        - agent.app_cert_files
        - agent.app_config
    ErrorResponse:
      type: object
      required: [error, message, details]
      properties:
        error:
          type: string
          example: validation_error
        message:
          type: string
        details:
          type: object
          description: Machine-readable error details. Empty object when there are no details.
          additionalProperties: true
...
    PageBase:
      type: object
      required: [total, offset, limit]
      properties:
        total:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
...
    JobProgress:
      type: object
      properties:
        completed:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
        current_target:
          type: string
          nullable: true
        current_scanner:
          type: string
          nullable: true
      additionalProperties: true
    JobResource:
      type: object
      description: Resource created by the job. For recompute, the AsyncJob itself is the resource and id equals JobEnvelope.id.
      required: [kind, id]
      properties:
        kind:
          $ref: '#/components/schemas/JobKind'
        id:
          type: integer
          description: Resource id for the job kind. For recompute this equals the API-visible job id.
    JobResult:
      type: object
      description: Completed-job result only. PENDING and RUNNING jobs must return null for JobEnvelope.result.
      properties:
        snapshot_id:
          type: integer
          nullable: true
        discovery_id:
          type: integer
          nullable: true
        updated_scores_count:
          type: integer
          description: Number of RiskScore rows updated by a completed recompute job.
          nullable: true
      additionalProperties: false
    JobEnvelope:
      type: object
      required: [id, kind, resource, status, progress, started_at, cancel_requested_at, finished_at, result, error]
      properties:
        id:
          type: integer
          description: API-visible asynchronous job id, unique across all job kinds.
        kind:
          $ref: '#/components/schemas/JobKind'
        resource:
          $ref: '#/components/schemas/JobResource'
        status:
          $ref: '#/components/schemas/JobStatus'
        progress:
          type: object
          allOf:
            - $ref: '#/components/schemas/JobProgress'
          nullable: true
          description: Always present. Null while PENDING or when no progress can be reported.
        started_at:
          type: string
          format: date-time
          nullable: true
        cancel_requested_at:
          type: string
          format: date-time
          nullable: true
          description: Set when cancellation has been requested but a running worker has not finished cleanup yet.
        finished_at:
          type: string
          format: date-time
          nullable: true
        result:
          type: object
          allOf:
            - $ref: '#/components/schemas/JobResult'
          nullable: true
          example:
            snapshot_id: 56
        error:
          type: string
          nullable: true
    JobEnvelopePage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/JobEnvelope'
    ScanJobCreate:
      type: object
      required: [target_ids, scanners]
      properties:
        target_ids:
          type: array
          minItems: 1
          items:
            type: integer
        scanners:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ScannerId'
    ScanRunLog:
      type: object
      required:
        - id
        - scan_job_id
        - target_id
        - target_label
        - scanner_kind
        - status
        - findings_count
      properties:
        id:
          type: integer
        scan_job_id:
          type: integer
        target_id:
          type: integer
        target_label:
          type: string
        scanner_kind:
          $ref: '#/components/schemas/ScannerId'
        status:
          $ref: '#/components/schemas/RunStatus'
        findings_count:
          type: integer
          minimum: 0
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          nullable: true
    ScanRunLogPage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ScanRunLog'
...